==> R/2024/1.0_read.Rmd <==
---
title: "Read"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: FALSE
    code_folding: show
    number_sections: TRUE
---

# SET-UP

```{r markdown, echo=F}
library(knitr)
opts_chunk$set(echo = T, collapse = T, fig.align = "center", fig.width = 9, fig.height = 6)
options(width = 200)
```    

```{r settings, message = F}
source('0.0_settings.R')
source("read.ziff.R")
```

# Landings
## ZIFF
```{r ziffsp, message = F}
sp_code <- find.species(my.species,'ziff')
kable(sp_code) 
```

```{r ziff, message = F}
#pending catch R update. Dowland from BD peches impossible. Copy on computer rename to start with "version_totale_" and change .txt to .csv

ziffR <- paste0(dir.rdat, "ziff.Rdata")

if(!file.exists(ziffR)){
    ziff <- read.ziff(sp_code,path=dir.ziff,year=my.year,language='en')
    
    # Species included:
    with(ziff, table(species,year))
    ziff <- ziff[ziff$species %in% c(250),]   # code 960: few heads in 1996  code 951: wrong whelk (clams main target..)
names(ziff)[19] <-  "nafo.sub"
names(ziff)[20] <-  "nafo"
    save(ziff, file = ziffR)
}else{
    load(ziffR)
}

# Creation time
df <- file.info(list.files(dir.rdat, full.names = T,pattern="ziff"))
df[which.max(df$mtime),'mtime']
```
### CHECKS {.tabset}

#### areas
Nafo 5Y and 5Ze are partially in the US.
```{r ziff areas}
# NAFO 5 = landings in Canadian EEZ (northern contingent)
kable(with(ziff, table(nafo,month)),caption = 'Number of records per month and NAFO area, entire database')     
```

#### NA weights
How many NA weights are there?
```{r ziff nas}
na <- ddply(ziff,c('year'),plyr::summarise,
            nacatch=length(catch[is.na(catch)]),
            nacatchland=length(catch[is.na(catch)&is.na(catch.landed)]),
            nacatchlandvalue=length(catch[is.na(catch)&is.na(catch.landed)&is.na(catch.value)]))
kable(na)
```

#### catch vs land weight
How different are both?
```{r ziff catchvsland}
dif <- (ziff$catch-ziff$catch.landed)/1000 #t
kable(round(quantile(dif[!is.na(dif)],c(0,0.025,0.25,0.5,0.75,0.975,1)),3),caption = 'Difference between catch.landed and catch (kg)')
```

#### 0 weights
How many 0 weights are there?
```{r ziff zeros}
with(ziff[!is.na(ziff$catch) & ziff$catch==0,],table(year))
```

#### Jan-1
Check that there is no abnormal quantity of landings on January first (possibly records for which data is unknown).
```{r ZIFF_jan1}
ggplot(ziff,aes(x=lubridate::yday(capt.date)))+geom_histogram(binwidth = 1)+
    scale_x_continuous(expand=c(0,0))+
    scale_y_continuous(expand=c(0,0))
```

#### NBPC
There are 3 special nbpc numbers that indicate landings that are not necessarily daily (e.g., monthly totals) or for which the data is only approximate. For instance, B00000 landings seems to be on the first of last of a certain month. For analyses of daily landings these might need to be excluded.
```{r ZIFF_npbc}
kable(with(ziff[ziff$npbc %in% c('000000', 'B00000', '190000'),], table(year,npbc)))  
```

#### Duplicates
There are rows that are 100% identical (same vessel, catch and landing date, trip, etc. AND even landed weight). Those are removed, but the effect is negligible.
```{r ZIFF_dups}
# Duplicates
dup <- data.frame(year=ziff$year,catch=ziff$catch,dup=duplicated(ziff))
dupd <- ddply(dup,c('year','dup'),plyr::summarise,catch=sum(catch,na.rm=TRUE))
ziff <- ziff[!dup$dup,]
ggplot(dupd,aes(x=year,y=catch/1000,fill=dup))+
    geom_bar(stat='identity',position='stack')+
    scale_x_continuous(expand=c(0,0))+
    scale_y_continuous(expand=c(0,0))+
    labs(x='Year',y='Landings (t)',fill='Duplicate')

```


## NAFO

```{r nafo, message = F}
nafoAR <- paste0(dir.rdat, "nafoA.Rdata")
nafoBR <- paste0(dir.rdat, "nafoB.Rdata")

if(!all(file.exists(c(nafoAR,nafoBR)))){
    ### NAFO database A (annual) ---------------------------------
    sp_code <- find.species(my.species,'nafoA');sp_code
    nafoA <- read.nafoA(paste0(dir.nafo,'statlandA.csv'),species=sp_code[1], year=my.year) 

    # Canadian caught mackerel outside of my.nafo
    with(nafoA[!grepl(paste0(my.nafo, collapse = "|"),nafoA$nafo)&
                 grepl('canada',nafoA$country,ignore.case = TRUE),] ,table(year,nafo))

    # only keep my.nafo; other areas are not from nothern contingent (winter months in US or Europe)
    nafoA <- nafoA[grepl(paste0(my.nafo, collapse = "|"),nafoA$nafo),] 
    table(nafoA$nafo)
    
    save(nafoA, file = nafoAR)
    
    ### NAFO data base B (monthly) ---------------------------------
    sp_code <- find.species(my.species,'nafoB');sp_code
    nafoB <- read.nafoB(path=dir.nafo,species=sp_code[1],year=my.year,overwrite=FALSE)  # overwrite=TRUE if new download   

    # Canadian caught mackerel outside of my.nafo
    with(nafoB[!grepl(paste0(my.nafo, collapse = "|"),nafoB$nafo)&
                 grepl('canada',nafoB$country,ignore.case = TRUE),] ,table(year,nafo))
    with(nafoB[!grepl(paste0(my.nafo, collapse = "|"),nafoB$nafo)&
                 grepl('canada',nafoB$country,ignore.case = TRUE),] ,table(month,nafo))
    
    # only keep my.nafo  (though 5Ze in summer; wouldn't they be northern contingent fish??)
    nafoB <- nafoB[grepl(paste0(my.nafo, collapse = "|"),nafoB$nafo),] 
    table(nafoB$nafo)
    
    # remove useless columns 
    nafoB <- nafoB[,c('year','month','country','nafo','gear','gear.cat','tonnage.class','prov','catch')]

    save(nafoB, file = nafoBR)
}else{
    load(nafoAR)
    load(nafoBR)
}

# Creation time
df <- file.info(list.files(dir.rdat, full.names = T,pattern="nafoB"))
df[which.max(df$mtime),'mtime']
```
### CHECK {.tabset}
Nafo B is used for landings prior to 1995.

#### compare
```{r NAFO_comparison}
ggplot()+
    geom_bar(data=nafoA,aes(x=year,y=catch,fill="Nafo A"),stat='identity',alpha=0.4)+
    geom_bar(data=nafoB,aes(x=year,y=catch,fill="Nafo B"),stat='identity',alpha=0.4)+
    geom_vline(xintercept=1985)+
    labs(fill='',y='Landings (t)',x='Year')+
    scale_x_continuous(expand=c(0,0))+ scale_y_continuous(expand=c(0,0))
```

## S52 Catch
```{r CATCHs52, message = F}
sel <- c('year','month','nafo','gear.cat','source','catch','country','prov')

lfR <- paste0(dir.rdat, "lf.Rdata")

    lf <- get.lf(species='maquereau',user=imlp.user,password=imlp.pass)
    
    # length: m to mm, consistent 5mm bins
    lf$length <-roundFish(lf$length*1000,5) # idem to carbio before
    
    # temporary correction for 2022: will be cleaned in the future!!!
    #lf[lf$sample.id==70450,'date'] <- as.Date("2022-09-02")
    #lf[lf$sample.id %in% c(69829,69830,69931),'sample.id'] <- 69829
    #lf[lf$sample.id %in% 70231,'source'] <- 91
    
    save(lf, file = lfR)

load(lfR)# weight landed for S52 license samples
sam <- lf[lf$year>=2022,]
sam <-  sam %>%  filter(source==91 | (source==99 & sample.id %in% c(73833:73838,73840))) %>% 
    dplyr::select('year','sample.id','month','wcatch','date','nafo','nbpc','source','gear','prov') %>%  distinct()

sam[is.na(sam$wcatch),'wcatch'] <- 300 # temporary 
nrow(sam)                               

catch.sam <- ddply(sam,c('year','month','nafo','gear','prov'),summarise,catch=sum(wcatch/1000))

catch.sam$source <- 'S52'
names(catch.sam)[which(names(catch.sam)=='gear')] <-'gear.cat'

pc<- catch.sam %>%  ggplot(aes(x=month, y=catch, fill=prov)) +geom_bar(stat="identity") +facet_wrap(~year) +
    scale_fill_viridis_d(name="") 
pc + labs(x="Mois | Month", y="Captures | Catch (t)")

kable(catch.sam  %>%  group_by(year) %>%  summarize(catchs52=sum(catch, na.rm=T)))

catch.s52<- catch.sam %>%  mutate(country="Canada", nafo.sub=nafo) %>%  dplyr::select(all_of(sel))

write.csv(catch.s52, paste0("../../csv/",max(my.year), "/S52_catch.csv"), row.names=F)


```



## CATCH = NAFOB + ZIFF + S52

```{r CATCH, message = F}
## bind both
ziff$source <-'ziff'
ziff$catch <- ziff$catch/1000            # kg to tonnes
ziff$country <- 'Canada'                 # consistency with nafo
ziff$prov <- ifelse(ziff$prov.land=='Unknown',as.character(ziff$prov.home),as.character(ziff$prov.land))
names(ziff)[which(names(ziff) =='gear_cat')] <- 'gear.cat'

nafoB$source <- 'nafo'                   # work with detailed B base




gulf24 <-  read_excel(paste0(dir.dat, "GLF2024_093_Mackerel Logbooks_2024.xlsx"), sheet=2)
gulf24 <- gulf24 %>%  dplyr::rename(year=`Landed Year`) %>%  mutate(catch= if_else(`Unit of Measure` == "GR", `Estimate Landing` /1000/1000,
                                                                 if_else(`Unit of Measure` == "KGS", `Estimate Landing` /1000,
                                                                         if_else(`Unit of Measure` == "LBS", `Estimate Landing` /2.205 /1000, NA))),
                                                   nafo=substring(`Det Unit Digits`,1,2),
                                                   source="pers.comm.",
                                                   prov="NB",
                                                   country="Canada",
                                        `Gear Code`= as.numeric(if_else(`Gear Entered` %in% c("FILET MAILLANT, DÉRIVANT"), 42,
                                                             if_else(`Gear Entered` %in% c("FILET MAILLANT, FIXE"), 41,
                                                                     if_else(`Gear Entered` %in% c("LIGNE À MAIN"), 59, 
                                                                             if_else(`Gear Entered` %in% c("41"), 41, `Gear Code`)))))  ,
                                        month=lubridate::month(`Date Landed`)
                                        ) %>%  dplyr::rename(engin=`Gear Code`) 
gulf24 %>%  group_by(`Gear Entered`, engin)  %>%  tally()
 ziff_gear <- read.csv(paste0(dir.dat,"/ziff_meta/ziff_gear.csv"), encoding="latin1") 

 
gulfok<-  left_join(gulf24, ziff_gear) %>%  dplyr::select(all_of(sel))

saveRDS(gulfok, paste0(dir.rdat,"GULF_2024.RDS"))


sel <- c('year','month','nafo','gear.cat','source','catch','country','prov')

qc = data.frame(year=2024, month=5:8, nafo=NA, gear.cat=NA, source="Bait", catch=c(458,12101,277,24279), country="Canada", prov="QC")

nl = data.frame(year=2024, month=5:8, nafo=NA, gear.cat=NA, source="Bait", catch=c(0,2621,1134,19454), country="Canada", prov="NL")

ziff<- ziff %>% mutate(flag = if_else(npbc %in% c("000000","B00000","190000") | as.numeric(as.character(npbc)) > 190000, 1,0)) 

catch <- bind_rows(ziff[ziff$year>=1995,c(sel,"nafo.sub","flag")],        # ziff from 1995 onwards
               nafoB[nafoB$year<1995,sel],       # nafoB before
               nafoB[nafoB$year>=1995 & nafoB$country!='Canada',sel], # foreign post 1994
                gulfok,
               qc %>%  mutate(catch=catch/1000),
               nl%>%  mutate(catch=catch/1000),
               catch.s52) # for data not yet in ziff
save(catch, file = paste0(dir.rdat, "catch.Rdata"))

plotCatch(catch %>%  mutate(dummy=1),x='year',y='catch', fill="dummy") +theme(legend.position = "none") +
    scale_x_continuous(breaks=c(seq(1970, max(my.year), 5)))
```
